+++ index.html
@@ -8,7 +8,9 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
   <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/data.js" type="module"></script>
+  <script src="/js/util.js" type="module"></script>
 </head>
 
 <body>
