+++ js/util.js
@@ -1,0 +1,12 @@
+// Функция для генерации случайных чисел в заданном диапазоне
+
+const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
+
+export {getRandomArrayElement, getRandomInt};
